Rename main file to index and remove package bin entry
The intention is that this is run via a script only now.
