+++ test/hanami-assets.test.ts
@@ -2,7 +2,7 @@ import fs from "fs-extra";
 import path from "path";
 import { globSync } from "glob";
 import crypto from "node:crypto";
-import * as assets from "../src/hanami-assets";
+import * as assets from "../src/index";
 
 const originalWorkingDir = process.cwd();
 const dest = path.resolve(__dirname, "..", "tmp", crypto.randomUUID());
